Added notification service

+++ src/app/tasks/components/task-list/task-list.component.ts
@@ -4,6 +4,7 @@ import { tasks } from '../../store/tasks.selector';
 import { TasksState } from '../../store/tasks.reducers';
 import { deleteTask } from '../../store/tasks.actions';
 import { Types } from '../../models/types';
+import { NotificationService } from "../../../services/notification.service";
 
 @Component({
   selector: 'app-task-list',
@@ -14,14 +15,18 @@ export class TaskListComponent{
   tasks$ = this.store.pipe(select(tasks));
   types = new Types().types;
   
-  constructor(private store: Store<TasksState>) {}
-
-  onDeleteTask(id: string) {
-    this.store.dispatch(deleteTask({id}));
-  }
+  constructor(private store: Store<TasksState>, public notificationService: NotificationService,) {}
 
   fieldLabel(type: string, field: any) {
     const currentType = this.types.filter(t => t.id === type);
     return currentType[0].fields[field].label;
+  }
+
+  onDeleteTask(id: string) {
+    this.notificationService.confirmation('', () => {
+      this.store.dispatch(deleteTask({id}));
+    },
+    'Are you sure to delete the task?',
+     () => {});
   }
 }
